LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/items/column-minimum-contribution-baseline-shim-vertical-lr-expected.txt


FAIL .grid-lanes 1 assert_equals:
<div class="grid-lanes" style="grid-template-columns: minmax(auto, 0px);">
  <div class="item1" data-offset-x="0" data-offset-y="0"></div>
  <div class="item2" data-offset-x="50" data-offset-y="50"></div>
  <div class="area" data-expected-width="75" data-expected-height="100"></div>
</div>
offsetLeft expected 50 but got 0
FAIL .grid-lanes 2 assert_equals:
<div class="grid-lanes" style="grid-template-columns: minmax(auto, 75px);">
  <div class="item1" data-offset-x="0" data-offset-y="0"></div>
  <div class="item2" data-offset-x="50" data-offset-y="50"></div>
  <div class="area" data-expected-width="75" data-expected-height="100"></div>
</div>
offsetLeft expected 50 but got 0
FAIL .grid-lanes 3 assert_equals:
<div class="grid-lanes" style="grid-template-columns: minmax(auto, 88px);">
  <div class="item1" data-offset-x="0" data-offset-y="0"></div>
  <div class="item2" data-offset-x="50" data-offset-y="50"></div>
  <div class="area" data-expected-width="88" data-expected-height="100"></div>
</div>
offsetLeft expected 50 but got 0
FAIL .grid-lanes 4 assert_equals:
<div class="grid-lanes" style="grid-template-columns: minmax(auto, 100px);">
  <div class="item1" data-offset-x="0" data-offset-y="0"></div>
  <div class="item2" data-offset-x="50" data-offset-y="50"></div>
  <div class="area" data-expected-width="100" data-expected-height="100"></div>
</div>
offsetLeft expected 50 but got 0
FAIL .grid-lanes 5 assert_equals:
<div class="grid-lanes" style="grid-template-columns: minmax(auto, 150px);">
  <div class="item1" data-offset-x="0" data-offset-y="0"></div>
  <div class="item2" data-offset-x="50" data-offset-y="50"></div>
  <div class="area" data-expected-width="100" data-expected-height="100"></div>
</div>
offsetLeft expected 50 but got 0
FAIL .grid-lanes 6 assert_equals:
<div class="grid-lanes" style="grid-template-columns: minmax(auto, auto);">
  <div class="item1" data-offset-x="0" data-offset-y="0"></div>
  <div class="item2" data-offset-x="50" data-offset-y="50"></div>
  <div class="area" data-expected-width="100" data-expected-height="100"></div>
</div>
offsetLeft expected 50 but got 0
